rename adhProposal in pcompass to avoid conflicts with core
