+++ src/pcompass/pcompass/static/js/Adhocracy.ts
@@ -42,7 +42,6 @@ import * as AdhPcompassModule from "./Packages/Pcompass/Module";
 import * as AdhPermissionsModule from "./Packages/Permissions/Module";
 import * as AdhPreliminaryNamesModule from "./Packages/PreliminaryNames/Module";
 import * as AdhProcessModule from "./Packages/Process/Module";
-import * as AdhProposalModule from "./Packages/Proposal/Module";
 import * as AdhRateModule from "./Packages/Rate/Module";
 import * as AdhResourceAreaModule from "./Packages/ResourceArea/Module";
 import * as AdhResourceWidgetsModule from "./Packages/ResourceWidgets/Module";
@@ -182,7 +181,6 @@ export var init = (config : AdhConfig.IService, metaApi) => {
     AdhPreliminaryNamesModule.register(angular);
     AdhProcessModule.register(angular);
     AdhProcessModule.register(angular);
-    AdhProposalModule.register(angular);
     AdhRateModule.register(angular);
     AdhResourceAreaModule.register(angular);
     AdhResourceWidgetsModule.register(angular);
