+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Module.ts
@@ -1,4 +1,5 @@
 import * as AdhPcompassContextModule from "./Context/Module";
+import * as AdhPcompassProposalModule from "./Proposal/Module";
 import * as AdhPcompassWorkbenchModule from "./Workbench/Module";
 
 
@@ -6,10 +7,12 @@ export var moduleName = "adhPcompass";
 
 export var register = (angular) => {
     AdhPcompassContextModule.register(angular);
+    AdhPcompassProposalModule.register(angular);
     AdhPcompassWorkbenchModule.register(angular);
 
     angular
         .module(moduleName, [
-            AdhPcompassContextModule.moduleName
+            AdhPcompassContextModule.moduleName,
+            AdhPcompassProposalModule.moduleName
         ]);
 };
