+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Proposal/Detail.html
