+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Proposal/ListItem.html
