+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Workbench/Module.ts
@@ -4,9 +4,10 @@ import * as AdhHttpModule from "../../Http/Module";
 import * as AdhMovingColumnsModule from "../../MovingColumns/Module";
 import * as AdhPermissionsModule from "../../Permissions/Module";
 import * as AdhProcessModule from "../../Process/Module";
-import * as AdhProposalModule from "../../Proposal/Module";
 import * as AdhResourceAreaModule from "../../ResourceArea/Module";
 import * as AdhTopLevelStateModule from "../../TopLevelState/Module";
+
+import * as AdhProposalModule from "../Proposal/Module";
 
 import * as AdhProcess from "../../Process/Process";
