+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Workbench/ProcessDetailColumn.html
@@ -29,7 +29,7 @@
             data-no-create-form="true"
             data-sorts="[{key: 'title', name: 'title'}]"
             data-empty-text="{{ 'TR__PROPOSAL_EMPTY_TEXT' | translate }}">
-            <adh-proposal-list-item data-path="{{element}}"></adh-proposal-list-item>
+            <adh-pcompass-proposal-list-item data-path="{{element}}"></adh-pcompass-proposal-list-item>
         </adh-listing>
     </div>
 </div>
