+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Workbench/ProposalCreateColumn.html
@@ -9,6 +9,6 @@
         </div>
     </div>
 
-    <adh-proposal-create data-ng-switch-when="body" >
-    </adh-proposal-create>
+    <adh-pcompass-proposal-create data-ng-switch-when="body" >
+    </adh-pcompass-proposal-create>
 </div>
