+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Workbench/ProposalDetailColumn.html
@@ -24,10 +24,10 @@
     <adh-recompile-on-change
         data-ng-switch-when="body"
         data-value="{{proposalUrl}}">
-        <adh-proposal-detail
+        <adh-pcompass-proposal-detail
             data-ng-if="proposalUrl"
             data-path="{{proposalUrl}}">
-        </adh-proposal-detail>
+        </adh-pcompass-proposal-detail>
     </adh-recompile-on-change>
 
     <adh-badge-assignment
