+++ src/pcompass/pcompass/static/js/Packages/Pcompass/Workbench/ProposalEditColumn.html
@@ -17,9 +17,9 @@
     <adh-recompile-on-change
             data-ng-switch-when="body"
             data-value="{{proposalUrl}}">
-        <adh-proposal-edit
+        <adh-pcompass-proposal-edit
             data-ng-if="proposalUrl"
             data-path="{{proposalUrl}}">
-        </adh-proposal-edit>
+        </adh-pcompass-proposal-edit>
     </adh-recompile-on-change>
 </div>
